refactor(cdc): use common datum ordering for string keys
